Implement first pass for drawing surface
For now we have a trivial test that can manipulate it's own drawing
buffer, then blit the entire buffer to the window in one step.
We have no need to rely on the drawing functions provided by xcb.

We had to partially guess the required data format for our drawing
buffer. We are currently successfully using BGRA bitmap data.