Get ship controls working
We have a reasonable starting point for how to handle movement in the
game.

Left and right controls rotate the ship. Thrust accelerates the ship in
the direction it's heading in. There is no reverse; the player has to
"flip and burn". However there is some level of enforced decelleration
when the ship is not under thrust.

Considering if left/right rotation should also be subject to retained
velocity. It may make the controls overly difficult.