Alter how we handle exiting
We have now fully gotten rid of the previous enum approach.
We also moved from using the 'q' key to using ESC instead. This makes
more sense; 'q' being directly next to 'w' was really asking for
trouble.

TODO:
We discovered that we're not handling the case where exit stems from the
user closing the window via the WM (with the X button).
To handle this, we need to do some platform-layer work, specifically
asking the WM to inform us of this event.
Ref:
https://github.com/GameTechDev/IntroductionToVulkan/blob/e8e384b3c5ccb71d13439ebd691ed6d297839f6c/Project/Common/OperatingSystem.cpp#L247