Simplify turning the ship
We only want to recalculate rotations when we need to.
Instead of rotating the acceleration vector and the lines that make up
the ship for every frame, from a base state, we now retain the rotation
across frames.

Further, we need to properly store the lines which make up the ship for
when we do collision detection.