Fix approach to line division
There were quite a few bugs and corner cases lurking in the line
division code. The original algorithm was taken almost directly from an
example of Cohen-Sutherland line-clipping found on wikipedia.
It was not intended to retain any clipped lines.

We are instead using it to divide our lines and then wrap them back into
screen space.

At this point, the algorithm seems to be fairly robust, at least for our
use-case. It does however feel like there should be ways to further
simplify/improve it as and when necessary.