Separate core game code from platform code
Further restructure the code to allow for a clean separation between the
core and platform layers.

We now compile the platform and core separately, and dynamically load
the core at runtime.

The two were still somewhat coupled at least by shared usage of the
arena code. However, we don't really need arena's in the platform layer,
so we moved to taking a more raw approach there.

The platform layer simply gets a big chunk of memory from the OS and
determines the key partition sizes - for our draw buffer, persistent
and temporary memory.

The draw buffer gets stored at the base of this chunk.
The platform layer then simply hands off control to the core game code;
for initialisation, passing only a pointer to the chunks' remaining
memory, the sizes for each of the persistent and temporary partitions
and a pointer to the draw buffer.

The core code determines the rest of the memory layout.
To allow it to keep a consistent state, the first thing it does is cast
the raw pointer to a game_state struct.

The game state is initialised to create separate arena's from which all
game allocations are sourced.
Aside from that, it may be used to store any other necessary state.

Because ultimately all of the data is stored by the platform layer, we
are now free to "hot-load" the core at any time and it can simply pick
up where it left off (other than when breaking changes are made).
This approach should further facilitate some interesting functionality
down the line.