Fix shared object load path
The core game code shared object is compiled in the same directory as
the main binary. It's load path was based only on it's filename, which
means the game would only successfully run when launched from the build
directory.

Now the .so load path is determined based on the path for the main
binary.