Implement hot-loading
When we compile our core code, we now do so to a temporay build_ file
which is immediately moved to our core .so upon completion. This
prevents a race condition where we may try to hot-load before the new
library has been fully written out.

Currently the modify time is checked every frame to determine if the
core has been recompiled and if so, we reload. Within reason, this gives
us live code editing.

Further, we first copy the .so and load this file during runtime instead
of the original. It has been noted that this may be necessary on some
platforms, either due to the OS or perhaps a debugger, where the file is
locked, preventing overwriting.
On linux however, it appears that overwriting may not ba a problem, so
we are undecided at present whether or not to bother with the copying.