Separate line splitting from drawing
To be able to handle collision detection properly we need to first have
the lines for all objects be split and placed in their final onscreen
positions, crucially accounting for any wrapping.

An alternative approach could be that as an object overlaps a screen
boundary, rather than splitting and wrapping lines, we could simply
duplicate the object and wrap the position of the whole duplicate.
Perhaps we will explore this later.